requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
python-dotenv==1.0.*
Flask-Bcrypt==1.0.*
Flask-JWT-Extended==4.5.*
PyJWT==2.9.*
Flask-Cors==3.0.*
flask-marshmallow==1.2.*
marshmallow-sqlalchemy==1.1.*
flasgger==0.9.7.*
pytest==8.3.*
pytest-flask==1.3.*
Flask-Migrate==4.0.*
alembic==1.14.*
